requirements.txt
# Smart Bin — Python dependencies
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0

# Evaluation & metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
